Renamed utils's enableExtension to enableExtensionHelpers
