+++ src/modules/dashboard/sagas/utils/enableExtensionHelpers.ts
