+++ src/modules/dashboard/sagas/utils/index.ts
@@ -7,4 +7,4 @@ export {
   removeOldExtensionClients,
   setupEnablingGroupTransactions,
   Channel,
-} from './enableExtension';
+} from './enableExtensionHelpers';
